+++ zf_core/include/zc/mem/bit_vector.h
@@ -9,7 +9,7 @@ namespace zf {
         c_bit_vector() = default;
 
         c_bit_vector(const c_array<t_byte> bytes, const size_t bit_cnt) : m_bytes(bytes), m_bit_cnt(bit_cnt) {
-            ZF_ASSERT(bytes.Len() == BitsToBytes(bit_cnt));
+            ZF_ASSERT(bytes.Len() == static_cast<int>(BitsToBytes(bit_cnt)));
         }
 
         [[nodiscard]] bool Init(c_mem_arena& mem_arena, const size_t bit_cnt);
@@ -23,17 +23,17 @@ namespace zf {
 
         void SetBit(const size_t index) const {
             ZF_ASSERT(index < m_bit_cnt);
-            m_bytes[index / 8] |= BitMask(index);
+            m_bytes[static_cast<int>(index / 8)] |= BitMask(index);
         }
 
         void UnsetBit(const size_t index) const {
             ZF_ASSERT(index < m_bit_cnt);
-            m_bytes[index / 8] &= ~BitMask(index);
+            m_bytes[static_cast<int>(index / 8)] &= ~BitMask(index);
         }
 
         bool IsBitSet(const size_t index) const {
             ZF_ASSERT(index < m_bit_cnt);
-            return m_bytes[index / 8] & BitMask(index);
+            return m_bytes[static_cast<int>(index / 8)] & BitMask(index);
         }
 
     private:
